pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["autoskill", "integrations", "tests", "core", "llm", "config", "templates", "utils"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.py", "*.yaml", "*.json", "*.md", "*.safetensors", "*.txt"]

[project]
name = "autoskill-ai"
version = "1.0.6"
description = "AI agent self-evolution plugin supporting dynamic skill generation, execution and management"
authors = [
    { name = "AutoSkill Team", email = "dataanswer@163.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.1.0",
    "langchain_openai>=0.1.0",
]
ml = [
    "numpy>=1.21.2",
    "pandas>=1.3.3",
    "scikit-learn>=0.24.2",
]
nlp = [
    "nltk>=3.6.3",
    "transformers>=4.18.0",
]
deep-learning = [
    "tensorflow>=2.8.0",
    "torch>=1.9.0",
]
skill-fingerprint = [
    "sentence-transformers>=2.2.0",
]
all = [
    "autoskill-ai[langchain,ml,nlp,deep-learning,skill-fingerprint]",
]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0"
]

[project.scripts]
auto-skill = "autoskill.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''/(\n    \.git
    | \.venv
    | venvs
    | build
    | dist
)/'''

[tool.isort]
profile = "black"

[tool.flake8]
extend-ignore = ["E203", "W503"]
max-line-length = 88
